requirements.txt
crispy-bootstrap5==0.7.0
dj-database-url==0.5.0
Django==4.2.3
django-crispy-forms==1.13.0
django-filter==21.1
django-multiselectfield==0.1.12
django-storages==1.12.3
gunicorn==20.1.0
numpy==1.25.2
Pillow==9.0.0
psycopg2-binary==2.9.9
psycopg2-binary==2.9.9
playwright==1.42.0
python-dateutil==2.8.2
python-dotenv==0.21.1
reportlab==3.6.13
sqlparse==0.4.2
urllib3==1.26.8
whitenoise==5.3.0